requirements.txt
Flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
beautifulsoup4==4.12.2
scikit-learn==1.4.2
pandas==2.2.2
joblib==1.4.2
google-generativeai==0.5.2
